updated calcVolFlowsDP0DT0.m
Updated the codes to allow the calculations of the volumetric flow rates so that the axial pressure in the CSTRs remain equal to one another but varies in time. We call linprog.m inside the right hand side function to calculate the pseudo volumetric flow rates. The code has gotten much slower now due to the embedded LP solver.