Fixed Performance metrics calcs
Make sure that the flow is diverted to the product tanks correctly, by including the boolean on/off switches for the volumetric flow rates entering the product tanks. Also, made sure that the global bool(10) determines zero flow rates out from extract tank, when we do not recover the extract product.